ALU class updates

+++ src/Parts/ALU.java
@@ -4,10 +4,8 @@ public class ALU {
     public SREG sreg;
     private byte operand1;
     private byte operand2;
-    private byte result;
+    private short result;
     private short opcode;
-
-
 
 
 }
